Repair LineString bug
In old qua-kit, we combine all the LineString/MultiLineString objects in the scenario into one MultiLineString object on submission. If there are no such as objects in the scenario, we previously create an empty MultiLineString object, prompting an error when uploading again the submitted scenario. In the corrected version, we do not create the MultiLineString object if there are no LineString/MultiLineString objects in the scenario.